refactor: PillarBlock -> FacingPillarBlock
