+++ src/main/java/com/aetherteam/aether/block/AetherBlocks.java
@@ -4,6 +4,7 @@ import com.aetherteam.aether.Aether;
 import com.aetherteam.aether.block.construction.*;
 import com.aetherteam.aether.block.dungeon.*;
 import com.aetherteam.aether.block.miscellaneous.AetherFrostedIceBlock;
+import com.aetherteam.aether.block.miscellaneous.FacingPillarBlock;
 import com.aetherteam.aether.block.miscellaneous.UnstableObsidianBlock;
 import com.aetherteam.aether.blockentity.ChestMimicBlockEntity;
 import com.aetherteam.aether.blockentity.SkyrootBedBlockEntity;
@@ -157,8 +158,8 @@ public class AetherBlocks {
 
     public static final RegistryObject<RotatedPillarBlock> PILLAR = register("pillar",
             () -> new RotatedPillarBlock(Block.Properties.of(Material.STONE, MaterialColor.QUARTZ).strength(0.5F).sound(SoundType.METAL).requiresCorrectToolForDrops()));
-    public static final RegistryObject<PillarBlock> PILLAR_TOP = register("pillar_top",
-            () -> new PillarBlock(Block.Properties.of(Material.STONE, MaterialColor.QUARTZ).strength(0.5F).sound(SoundType.METAL).requiresCorrectToolForDrops()));
+    public static final RegistryObject<FacingPillarBlock> PILLAR_TOP = register("pillar_top",
+            () -> new FacingPillarBlock(Block.Properties.of(Material.STONE, MaterialColor.QUARTZ).strength(0.5F).sound(SoundType.METAL).requiresCorrectToolForDrops()));
 
     public static final RegistryObject<Block> PRESENT = register("present",
             () -> new Block(Block.Properties.of(Material.GRASS, MaterialColor.COLOR_GREEN).strength(0.1F).sound(SoundType.WOOL)));
